+++ pkg/splunk/splkcontroller/doc.go
@@ -18,4 +18,4 @@ Package controller is used to manipulate Kubernetes resources using its REST API
 This package has no depedencies outside of the standard go and kubernetes libraries,
 and the splunk.common package.
 */
-package controller
+package splkcontroller
