+++ pkg/splunk/splkcontroller/serviceaccount_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package controller
+package splkcontroller
 
 import (
 	"context"
